+++ flowman-core/src/main/scala/com/dimajix/flowman/config/FlowmanConf.scala
@@ -108,6 +108,10 @@ object FlowmanConf {
         .doc("Parallelism of mapping instantiation")
         .intConf
         .createWithDefault(1)
+    val EXECUTION_CHECK_PARALLELISM = buildConf("flowman.execution.check.parallelism")
+        .doc("Parallelism of data quality check execution")
+        .intConf
+        .createWithDefault(1)
     val EXECUTION_MAPPING_SCHEMA_CACHE = buildConf("flowman.execution.mapping.schemaCache")
         .doc("Cache schema information of mapping instances")
         .booleanConf