+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Version 0.29.1
+* github-278: Parallelize execution of data quality checks
+
+
 # Version 0.29.0 - 2022-11-08
 
 * github-260: Remove hive-storage-api from several plugins and lib